+++ src/files_proc/processor.rs
@@ -9,6 +9,7 @@ use crate::lib_utils::config::Config;
 use crate::lib_utils::errors::Vex2PdfError;
 use crate::pdf::generator::PdfGenerator;
 use crate::utils::{get_output_pdf_path, parse_vex_json, parse_vex_xml};
+use log::{error, info, warn};
 use std::collections::HashSet;
 use std::fs;
 use std::hash::Hash;
@@ -32,7 +33,7 @@ impl FileSearchProvider for DefaultFilesProcessor {
     type ErrType = Vex2PdfError;
     fn find_files(self) -> Result<Self::OkType, Self::ErrType> {
         // process map ignored pattern map
-        let ignored_patterns_map = (&self.config).file_types_to_process.as_ref();
+        let ignored_patterns_map = self.config.file_types_to_process.as_ref();
 
         // get the working path can be a file or folder
         let working_path = &self.config.working_path;
@@ -43,8 +44,8 @@ impl FileSearchProvider for DefaultFilesProcessor {
                 ignore_map
                     .iter()
                     .filter_map(|(k, v)| {
-                        if *v == false {
-                            println!(
+                        if !(*v) {
+                            info!(
                                 "Skipping {} files : deactivated by user",
                                 k.as_str_uppercase()
                             );
@@ -64,7 +65,7 @@ impl FileSearchProvider for DefaultFilesProcessor {
             ret.add_sup_file_ignore(working_path.to_path_buf(), &ignored_patterns)?;
         } else {
             // is a folder
-            println!(
+            info!(
                 "Scanning for BoM/Vex Files in {}",
                 &self.config.working_path.display()
             );
@@ -75,8 +76,8 @@ impl FileSearchProvider for DefaultFilesProcessor {
                 if path.is_file() {
                     if let Err(e) = ret.add_sup_file_ignore(path.to_path_buf(), &ignored_patterns) {
                         match e {
-                            Vex2PdfError::IgnoredByUser => println!("{e}"),
-                            _ => eprintln!("{e}"),
+                            Vex2PdfError::IgnoredByUser => info!("{e}"),
+                            _ => error!("{e}"),
                         };
                     }
                 }
@@ -85,14 +86,14 @@ impl FileSearchProvider for DefaultFilesProcessor {
 
         // inform over search results
         if ret.get_files_ref().is_empty() {
-            println!("No parseable files in selected path");
+            info!("No parseable files in selected path");
         } else {
-            println!(
+            info!(
                 "Found {} JSON files",
                 ret.get_file_count_by_type(InputFileType::JSON)
             );
-            println!(
-                "Fond {} XML files",
+            info!(
+                "Found {} XML files",
                 ret.get_file_count_by_type(InputFileType::XML)
             );
         }
@@ -115,41 +116,33 @@ impl<P: AsRef<Path> + Eq + Hash + Send + 'static> SingleFileProcProvider<P>
         file: BomFileIdentifier<P>,
         config: Arc<Config>,
     ) -> Result<(), Vex2PdfError> {
-        println!("Processing {}", file.get_path().as_ref().display());
+        info!("Processing {}", file.get_path().as_ref().display());
 
         // Get BoM Object
         let bom = match file.get_type() {
             InputFileType::XML => {
-                match parse_vex_xml(file.get_path()).map_err(|e| Vex2PdfError::Parse(e.to_string()))
-                {
-                    Ok(bom) => bom,
-                    Err(e) => return Err(e),
-                }
-            }
+                parse_vex_xml(file.get_path()).map_err(|e| Vex2PdfError::Parse(e.to_string()))?
+            },
             InputFileType::JSON => {
-                match parse_vex_json(file.get_path())
-                    .map_err(|e| Vex2PdfError::Parse(e.to_string()))
-                {
-                    Ok(bom) => bom,
-                    Err(e) => return Err(e),
-                }
-            }
-            InputFileType::UNSUPPORTED => return Err(Vex2PdfError::UnsupportedFileType),
+                parse_vex_json(file.get_path())
+                    .map_err(|e| Vex2PdfError::Parse(e.to_string()))?
+            },
+            InputFileType::UNSUPPORTED => return Err(Vex2PdfError::UnsupportedFileType)
         };
 
         // Generate output PDF path with same base name
         let generator = PdfGenerator::new(Arc::clone(&config));
 
-        println!("Generating PDF:  {}", file.get_path().as_ref().display());
+        info!("Generating PDF:  {}", file.get_path().as_ref().display());
 
         // FIXME consider if output path is ever handled here
         // Generate the PDF
         let output_path =
             get_output_pdf_path(Some(config.output_dir.as_path()), file.get_path().as_ref())?;
         match generator.generate_pdf(&bom, &output_path) {
-            Ok(_) => println!("Successfully generated PDF: {}", output_path.display()),
+            Ok(_) => info!("Successfully generated PDF: {}", output_path.display()),
             Err(e) => {
-                println!(
+                warn!(
                     "Failed to generate PDF for {}: {}",
                     output_path.display(),
                     e
@@ -178,12 +171,12 @@ impl<P: AsRef<Path> + Eq + Hash + Send + 'static> MultipleFilesProcProvider<P>
         let file_count = self.files.get_file_count();
 
         for file in self.files {
-            let single_file_proc = DefaultSingleFileProcessor::default();
+            let single_file_proc = DefaultSingleFileProcessor;
 
             let config_clone = Arc::clone(&config);
             pool.execute(move || {
                 if let Err(e) = single_file_proc.process_single_file(file, config_clone) {
-                    eprintln!("{e}");
+                    error!("{e}");
                 }
             })
             .expect(
@@ -194,7 +187,7 @@ impl<P: AsRef<Path> + Eq + Hash + Send + 'static> MultipleFilesProcProvider<P>
         drop(pool); // dropping here to show information message after worker status messages
                     // pool drops gracefully and cleans up here blocking until all jobs are finished
 
-        println!("Processed {file_count} files");
+        info!("Processed {file_count} files");
 
         Ok(())
     }