+++ src/main.rs
@@ -25,18 +25,47 @@
 //! No extra configuration is required
 //! See the README for more details.
 
+use log::error;
+use std::io::Write;
 use std::process;
 use vex2pdf::lib_utils::config::Config;
 
 fn main() {
+    env_logger::Builder::from_env(env_logger::Env::default().default_filter_or("info"))
+        .format(|buf, record| {
+            // Write errors and warnings to stderr, info and debug to stdout
+            match record.level() {
+                log::Level::Error | log::Level::Warn => {
+                    writeln!(
+                        &mut std::io::stderr(),
+                        "[{} {}] {}",
+                        buf.timestamp(),
+                        record.level(),
+                        record.args()
+                    )
+                }
+                _ => {
+                    writeln!(
+                        buf,
+                        "[{} {}] {}",
+                        buf.timestamp(),
+                        record.level(),
+                        record.args()
+                    )
+                }
+            }
+        })
+        .target(env_logger::Target::Stdout)
+        .init();
+
     let config = Config::build().unwrap_or_else(|err| {
-        eprintln!("Problem setting up working environment:");
-        eprintln!("{}", { err });
+        error!("Problem setting up working environment:");
+        error!("{}", { err });
         process::exit(1);
     });
 
     if let Err(e) = vex2pdf::run(config) {
-        eprintln!("Application error: {e}");
+        error!("Application error: {e}");
         process::exit(1);
     }
 }